issue_14

+++ app/Http/Controllers/PedidoController.php
@@ -17,6 +17,7 @@ use App\Jobs\DiagramacaoJob;
 use App\Jobs\ImpressaoJob;
 use App\Jobs\AcabamentoJob;
 use App\Jobs\FinalizarJob;
+use Illuminate\Validation\Rule;
 
 class PedidoController extends Controller
 {
@@ -27,11 +28,20 @@ class PedidoController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('servidor');
         $request->validate([
             'busca_data' => 'required_if:filtro_busca,data|dateformat:d/m/Y|nullable',
+            'busca_tipo' => ['nullable',Rule::in(Pedido::tipoOptions())],
+            'busca_status' => ['nullable',Rule::in(Pedido::status)],
         ]);
+
+        if($request->busca_status != ''){
+            $query = Pedido::currentStatus("{$request->busca_status}")->join('users', 'users.id', '=', 'pedidos.user_id')->orderBy('pedidos.created_at', 'desc')->select('pedidos.*'); 
+        }
+        else{
+            $query = Pedido::join('users', 'users.id', '=', 'pedidos.user_id')->orderBy('pedidos.created_at', 'desc')->select('pedidos.*'); 
+        }
         
-        $query = Pedido::join('users', 'users.id', '=', 'pedidos.user_id')->orderBy('pedidos.created_at', 'desc')->select('pedidos.*'); 
         if($request->busca != ''){
             $query->where(function($query) use($request){
                 $query->orWhere('users.name', 'LIKE', "%$request->busca%");
@@ -55,14 +65,62 @@ class PedidoController extends Controller
         return view('pedidos.index')->with('pedidos',$pedidos);
     }
 
-    public function meusPedidos(){
-        $pedidos = Pedido::where('user_id', auth()->user()->id)->paginate(20);
+    public function meusPedidos(Request $request){
+        $this->authorize('logado');
+        $request->validate([
+            'busca_tipo' => ['nullable',Rule::in(Pedido::tipoOptions())],
+            'busca_status' => ['nullable',Rule::in(Pedido::status)],
+        ]);
+
+        if($request->busca_status != ''){
+            $query = Pedido::currentStatus("{$request->busca_status}")->where('user_id', auth()->user()->id)->orderBy('created_at', 'desc'); 
+        }
+        else{
+            $query = Pedido::where('user_id', auth()->user()->id)->orderBy('created_at', 'desc'); 
+        }
+        if($request->busca_tipo != ''){
+            $query->where('tipo','=', $request->busca_tipo);
+        }
+        
+        $pedidos = $query->paginate(20);
+        
+        if ($pedidos->count() == null and $request->busca != '') {
+            $request->session()->flash('alert-danger', 'Não há registros!');
+        }
         return view('pedidos.meus_pedidos')->with('pedidos',$pedidos);
     }
 
-    public function autorizacaoPedidos(){
-        $pedidos = Pedido::currentStatus('Autorização');
-        $pedidos = $pedidos->where('responsavel_centro_despesa', auth()->user()->codpes)->paginate(20);
+    public function autorizacaoPedidos(Request $request){
+        $this->authorize('logado');
+        $request->validate([
+            'busca_tipo' => ['nullable',Rule::in(Pedido::tipoOptions())],
+            'busca' => 'nullable',
+            'busca_status' => ['nullable',Rule::in(Pedido::status)],
+        ]);
+
+        if($request->busca_status != ''){
+            $query = Pedido::currentStatus("{$request->busca_status}")->join('users', 'users.id', '=', 'pedidos.user_id')->where('responsavel_centro_despesa', auth()->user()->codpes)->orderBy('pedidos.created_at', 'desc')->select('pedidos.*'); 
+        }
+        else{
+            $query = Pedido::currentStatus("Autorização")->join('users', 'users.id', '=', 'pedidos.user_id')->where('responsavel_centro_despesa', auth()->user()->codpes)->orderBy('pedidos.created_at', 'desc')->select('pedidos.*'); 
+        }
+        
+        if($request->busca != ''){
+            $query->where(function($query) use($request){
+                $query->orWhere('users.name', 'LIKE', "%$request->busca%");
+                $query->orWhere('users.codpes', '=', "$request->busca");
+                $query->orWhere('pedidos.descricao', 'LIKE', "%$request->busca%");
+            });
+        }
+        if($request->busca_tipo != ''){
+            $query->where('tipo','=', $request->busca_tipo);
+        }
+        
+        $pedidos = $query->paginate(20);
+        
+        if ($pedidos->count() == null and $request->busca != '') {
+            $request->session()->flash('alert-danger', 'Não há registros!');
+        }
         return view('pedidos.autorizacao_pedidos')->with('pedidos',$pedidos);
     }
 
@@ -73,6 +131,7 @@ class PedidoController extends Controller
      */
     public function create()
     {
+        $this->authorize('logado');
         $pedido = new Pedido;
         return view('pedidos.create')->with('pedido', $pedido);
     }
@@ -85,6 +144,7 @@ class PedidoController extends Controller
      */
     public function store(PedidoRequest $request)
     {
+        $this->authorize('logado');
         $validated = $request->validated();
         $pedido = Pedido::create($validated);
         $pedido->setStatus('Em Elaboração');
@@ -100,6 +160,7 @@ class PedidoController extends Controller
      */
     public function show(Pedido $pedido, PedidoStepper $stepper)
     {
+        $this->authorize('owner.pedido',$pedido);
         $stepper->setCurrentStepName($pedido->latestStatus());
 
         return view('pedidos.show', [
@@ -116,6 +177,7 @@ class PedidoController extends Controller
      */
     public function edit(Pedido $pedido)
     {
+        $this->authorize('owner.pedido',$pedido);
         return view('pedidos.edit')->with('pedido', $pedido);
     }
 
@@ -128,6 +190,7 @@ class PedidoController extends Controller
      */
     public function update(PedidoRequest $request, Pedido $pedido)
     {
+        $this->authorize('owner.pedido',$pedido);
         $validated = $request->validated();
         $pedido->update($validated);
         return redirect("/pedidos/$pedido->id");
@@ -141,18 +204,24 @@ class PedidoController extends Controller
      */
     public function destroy(Pedido $pedido)
     {
-        $pedido->orcamentos()->delete();
-        $files = $pedido->files;
-        foreach($files as $file){
-            Storage::delete($file->path);
+        $this->authorize('owner.pedido',$pedido);
+        if($pedido->status == 'Em Elaboração'){
+            $pedido->orcamentos()->delete();
+            $files = $pedido->files;
+            foreach($files as $file){
+                Storage::delete($file->path);
+            }
+            $pedido->files()->delete();
+            $pedido->delete();
+            return redirect('/pedidos');
         }
-        $pedido->files()->delete();
-        $pedido->delete();
-        return redirect('/pedidos');
     }
 
     //Funções de Status
+
+    //Função que envia pedido para liberação do grupo AUTORIZADOR
     public function enviarAnalise(Pedido $pedido, Request $request){
+        $this->authorize('owner.pedido', $pedido);
         $pedido->setStatus('Em Análise', $request->reason);
         foreach(explode(',', trim(env('AUTORIZADOR'))) as $codpes){
             if(Pessoa::emailusp($codpes)){
@@ -162,7 +231,11 @@ class PedidoController extends Controller
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função que envia pedido, caso liberado, para o grupo EDITORA e GRÁFICA
+    //para que façam o orçamento do pedido
+    //caso rejeitado, status volta para 'Em Elaboração' e é devolvido para o usuário
     public function enviarOrcamento(Pedido $pedido, Request $request){
+        $this->authorize('servidor');
         if($request->button == 'orcamento'){
             $pedido->setStatus('Orçamento', $request->reason);
             foreach(explode(',', trim(env('EDITORA'))) as $codpes){
@@ -183,15 +256,25 @@ class PedidoController extends Controller
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função ativada após feitura do orçamento, em que a editora ou gráfica
+    //envia o orçamento para o usuário e para o responsável do centro de despesa para que este autorize
     public function autorizacao(Pedido $pedido, Request $request){
+        $this->authorize('servidor');
         $pedido->setStatus('Autorização', $request->reason);
         if(Pessoa::emailusp($pedido->responsavel_centro_despesa)){
-            AutorizacaoJob::dispatch($pedido);
+            AutorizacaoJob::dispatch($pedido, $pedido->responsavel_centro_despesa);
+        }
+        if(Pessoa::emailusp($pedido->user->codpes)){
+            AutorizacaoJob::dispatch($pedido, $pedido->user->codpes);
         }
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função que trabalha com o resultado da autorização do Centro de Despesa e encaminha
+    //para os próximos passos do sistema (indo para a Editora ou para a Gráfica)
+    //também pode ocorrer do Centro de Despesa não liberar, então retorna para status 'Em Elaboração'
     public function enviarAutorizacao(Pedido $pedido, Request $request){
+        $this->authorize('owner.pedido', $pedido);
         if($request->button == 'autorizado'){
             if($pedido->tipo == 'Diagramação' or $pedido->tipo == 'Diagramação + Impressão'){
                 $pedido->setStatus('Diagramação', $request->reason);
@@ -217,7 +300,9 @@ class PedidoController extends Controller
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função que encaminha para a Gráfica assim que termina o status de Diagramação
     public function impressao(Pedido $pedido, Request $request){
+        $this->authorize('editora');
         $pedido->setStatus('Impressão', $request->reason);
         foreach(explode(',', trim(env('GRAFICA'))) as $codpes){
             if(Pessoa::emailusp($codpes)){
@@ -227,13 +312,17 @@ class PedidoController extends Controller
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função que avisa o usuário do acabamento do pedido na gráfica
     public function acabamento(Pedido $pedido, Request $request){
+        $this->authorize('grafica');
         $pedido->setStatus('Acabamento', $request->reason);
         AcabamentoJob::dispatch($pedido);
         return redirect("/pedidos/{$pedido->id}");
     }
 
+    //Função que avisa o usuário da finalização do pedido
     public function finalizar(Pedido $pedido, Request $request){
+        $this->authorize('servidor');
         $pedido->setStatus('Finalizado', $request->reason);
         FinalizarJob::dispatch($pedido);
         return redirect("/pedidos/{$pedido->id}");
@@ -262,6 +351,7 @@ class PedidoController extends Controller
         } 
     }
 
+    //Função para link de acesso temporário do arquivo do pedido
     public function acesso_autorizado(Request $request)
     {
         if ($request->hasValidSignature()) {
